posts from all categories/subcategories are now displayed when "Naslovna" is active, using CustomPostsSection/ cache logic also works for "Naslovna" and all categories/subcategories
Posts on "Naslovna" are displayed horizontally and can be scrollable, with the category names above them. CustomPostsSection is used to design posts coming from the api. Posts are stored in asyncStorage  after the first get (only title and image, because of the limit storage). When the post is clicked, the other details of it are fetched and displayed. This works for every category and subcategory. If cache is not reachable for some reason, system tries to fetch posts again. This logic allows the user to use the app without internet (after the first fetch is done) and without needing to always fetch posts on open (this takes too much time).