favorites page added with refresh logic
On favorites page, category of the saved post is saved as well as posts in horizontal scroll view. Whole category of saved posts can be deleted and one post can also be deleted. There is one more alert to prevent from accidental removals. If that happens, the page is auto refreshed and posts are deleted immediately.