notifications with OneSignal work well/ they are stored and displayed on notifications.tsx
When a notification is received by the user who is outside the app, and he clicks on the notification, it opens up the app on the post that was posted from that notification. If the user is inside the app when a new notification appears it nicely pops up on the menu.  All notifications are displayed on the notifications screen and stored in local storage. Notifications can also be marked read/unread. There is a new useOneSignalDeepLinks hook and the rest are helper modules for _layout.tsx with OneSignal integration.