post-details updated to have a separate logic for displaying posts from a notification for now (will be changed)/ also search works on post-details now
Post-details now checks if the post has come from a notification or local storage for now (should be simplified later). If the user tries to press search when on post details screen it redirects him to a search input field that is used on index screen, and the search works the same from there.