ads added and they work across the whole app
On "Naslovna" ads are displayed below every category, and there is another ad that pops up after a certain period has passed above the footer menu. On other categories there is one add at the bottom. On all other screens there also exists an ad that appears after a certain period has passed on that screen. The ad can be closed, but it appears again after a while. For the ads on index, they can be removed but they appear again when the application reopens. For now there are 5 different ads that are displayed randomly across the whole app, but that should be upgraded and more ads should be added.